# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  docsResponse.ImageUpload200:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_image_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.ImageUpload400:
    properties:
      errorCode:
        enum:
        - BAD_REQUEST
        type: string
      fields:
        items:
          $ref: '#/definitions/docsResponse.imageErrorField'
        type: array
      isSuccess:
        example: false
        type: boolean
      message:
        example: Bad request or validation error
        type: string
    type: object
...
  docsResponse.Response401:
    properties:
      errorCode:
        enum:
        - UNAUTHORIZED
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Unauthorized
        type: string
    type: object
...
  docsResponse.Response413:
    properties:
      errorCode:
        enum:
        - REQUEST_TOO_LARGE
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Request entity too large
        type: string
    type: object
  docsResponse.Response500:
    properties:
      errorCode:
        enum:
        - SERVER_ERROR
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Internal server error
        type: string
    type: object
...
  docsResponse.imageErrorField:
    properties:
      errorCode:
        example: INVALID_FORMAT
        type: string
      field:
        example: images
        type: string
    type: object
...
  haircompany-shop-rest_internal_modules_v1_image_dto.ResponseDTO:
    properties:
      image:
        type: string
    type: object
host: https://{host}:{port}
info:
  contact:
    email: x3.na.tri@gmail.com
    name: API Support
  description: REST API для магазина бренда Hair Company
  title: Hair Company Shop API
  version: "1.0"
paths:
...
  /api/v1/image/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload one or more images with a specific imageType. Requires JWT
        authentication.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Type of image (e.g. category)
        in: formData
        name: imageType
        required: true
        type: string
      - description: Image files to upload (multiple allowed)
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Uploaded images info
          schema:
            items:
              $ref: '#/definitions/docsResponse.ImageUpload200'
            type: array
        "400":
          description: Bad Request or Validation Error
          schema:
            $ref: '#/definitions/docsResponse.ImageUpload400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docsResponse.Response401'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/docsResponse.Response413'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      security:
      - BearerAuth: []
      summary: Upload images
      tags:
      - Image
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"